opendbc changes (read description)
This gives opendbc the ability to decode Toyota's light sensor value, this is necessary for this feature to work. I have already submitted a PR to comma and it may get merged.

This will be no longer necessary once comma merges my PR (provided you will be willing to update your opendbc).